move screens dir
