+++ src/navigators/AppNavigator.js
@@ -2,15 +2,27 @@ import React from "react"
 import { addNavigationHelpers, StackNavigator } from "react-navigation"
 import { connect } from "react-redux"
 
-import LoginScreen from "../components/screens/LoginScreen"
-import RegisterScreen from "../components/screens/RegisterScreen"
+import LoginScreen from "../screens/LoginScreen"
+import RegisterScreen from "../screens/RegisterScreen"
 
 import MainDrawer from "./MainDrawer"
 
 export const AppNavigator = StackNavigator({
-  Login: { screen: LoginScreen },
-  Register: { screen: RegisterScreen },
-  MainDrawer: { screen: MainDrawer },
+  Login: {
+    screen: LoginScreen,
+    navigationOptions: {
+      title: "Login",
+    },
+  },
+  Register: {
+    screen: RegisterScreen,
+    navigationOptions: {
+      title: "Register",
+    },
+  },
+  MainDrawer: {
+    screen: MainDrawer,
+  },
 })
 
 const AppWithNavigationState = ({ dispatch, nav }) => (
