+++ src/navigators/ProductTab.js
@@ -1,8 +1,8 @@
 import React from "react"
 import { TabNavigator } from "react-navigation"
 
-import IndexProductScreen from "../components/screens/products/IndexProductScreen"
-import AddProductScreen from "../components/screens/products/AddProductScreen"
+import IndexProductScreen from "../screens/products/IndexProductScreen"
+import AddProductScreen from "../screens/products/AddProductScreen"
 
 const ProductTab = TabNavigator({
   IndexProduct: {
